# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fleet Inventory Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MachineFlagSeverity = typeof MachineFlagSeverity[keyof typeof MachineFlagSeverity];


export const MachineFlagSeverity = {
  danger: 'danger',
  warn: 'warn',
  ok: 'ok',
} as const;

export interface MachineFlag {
  label: string;
  severity: MachineFlagSeverity;
}

/**
 * @nullable
 */
export type MachineData = { [key: string]: unknown } | null;

export interface Machine {
  machine_id: string;
  hostname: string;
  /** @nullable */
  site?: string | null;
  last_seen: string;
  /** @nullable */
  manufacturer?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  cpu?: string | null;
  /** @nullable */
  total_ram_gb?: number | null;
  /** @nullable */
  ram_type?: string | null;
  /** @nullable */
  gpu1_model?: string | null;
  /** @nullable */
  os?: string | null;
  /** @nullable */
  primary_ip?: string | null;
  /** @nullable */
  data?: MachineData;
  flags: MachineFlag[];
}

/**
 * @nullable
 */
export type MachineReportData = { [key: string]: unknown } | null;

export interface MachineReport {
  machine_id: string;
  hostname: string;
  /** @nullable */
  site?: string | null;
  /** @nullable */
  manufacturer?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  cpu?: string | null;
  /** @nullable */
  total_ram_gb?: number | null;
  /** @nullable */
  ram_type?: string | null;
  /** @nullable */
  gpu1_model?: string | null;
  /** @nullable */
  os?: string | null;
  /** @nullable */
  primary_ip?: string | null;
  /** @nullable */
  data?: MachineReportData;
}

export interface DashboardStats {
  total_machines: number;
  flagged_machines: number;
  site_count: number;
  danger_count: number;
  warn_count: number;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface UserSession {
  id: number;
  username: string;
  role: string;
}

export interface User {
  id: number;
  username: string;
  role: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  viewer: 'viewer',
} as const;

export interface UserInput {
  username: string;
  password: string;
  role: UserInputRole;
}

export interface PasswordReset {
  password: string;
}

export interface UpdateMachineSiteInput {
  /** @nullable */
  site: string | null;
}

export interface SubnetRule {
  id: number;
  cidr: string;
  site: string;
  created_at: string;
}

export interface SubnetRuleInput {
  /** IPv4 CIDR, e.g. 10.1.0.0/16 */
  cidr: string;
  site: string;
}

export type ListMachinesParams = {
/**
 * Free-text search across host/IP/CPU/GPU/model/OS
 * @nullable
 */
search?: string | null;
/**
 * Filter by site
 * @nullable
 */
site?: string | null;
/**
 * Filter flagged machines only
 * @nullable
 */
flagged?: boolean | null;
};