# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fleet Inventory Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: machines
    description: Machine inventory
  - name: auth
    description: Authentication
  - name: users
    description: User management
  - name: stats
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /machines:
    get:
      operationId: listMachines
      tags: [machines]
      summary: List all machines
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
          description: Free-text search across host/IP/CPU/GPU/model/OS
        - name: site
          in: query
          schema:
            type: ["string", "null"]
          description: Filter by site
        - name: flagged
          in: query
          schema:
            type: ["boolean", "null"]
          description: Filter flagged machines only
      responses:
        "200":
          description: List of machines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Machine"
        "401":
          description: Unauthorized

  /machines/{machine_id}:
    get:
      operationId: getMachine
      tags: [machines]
      summary: Get machine detail
      parameters:
        - name: machine_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Machine detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    patch:
      operationId: updateMachineSite
      tags: [machines]
      summary: Update a machine's site (admin)
      parameters:
        - name: machine_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMachineSiteInput"
      responses:
        "200":
          description: Updated machine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteMachine
      tags: [machines]
      summary: Delete a machine
      parameters:
        - name: machine_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

  /report:
    post:
      operationId: reportMachine
      tags: [machines]
      summary: Ingest machine hardware report
      description: Protected by Bearer token (INGEST_TOKEN env var)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MachineReport"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
        "401":
          description: Unauthorized

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized

  /sites:
    get:
      operationId: listSites
      tags: [stats]
      summary: List unique sites
      responses:
        "200":
          description: List of site names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "401":
          description: Unauthorized

  /subnets:
    get:
      operationId: listSubnets
      tags: [subnets]
      summary: List subnet-to-site mapping rules
      responses:
        "200":
          description: List of subnet rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubnetRule"
        "401":
          description: Unauthorized
    post:
      operationId: createSubnet
      tags: [subnets]
      summary: Create a subnet-to-site mapping rule (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubnetRuleInput"
      responses:
        "200":
          description: Created subnet rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubnetRule"
        "400":
          description: Invalid CIDR
        "401":
          description: Unauthorized
        "403":
          description: Forbidden

  /subnets/{id}:
    delete:
      operationId: deleteSubnet
      tags: [subnets]
      summary: Delete a subnet-to-site mapping rule (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

  /login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSession"
        "401":
          description: Invalid credentials

  /logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current session
      responses:
        "200":
          description: Logged out

  /me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSession"
        "401":
          description: Not authenticated

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden

  /users/{user_id}:
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    patch:
      operationId: updateUserPassword
      tags: [users]
      summary: Reset a user password (admin only)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordReset"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    MachineFlag:
      type: object
      required: [label, severity]
      properties:
        label:
          type: string
        severity:
          type: string
          enum: [danger, warn, ok]

    Machine:
      type: object
      required:
        - machine_id
        - hostname
        - last_seen
        - flags
      properties:
        machine_id:
          type: string
        hostname:
          type: string
        site:
          type: ["string", "null"]
        last_seen:
          type: string
          format: date-time
        manufacturer:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        cpu:
          type: ["string", "null"]
        total_ram_gb:
          type: ["integer", "null"]
        ram_type:
          type: ["string", "null"]
        gpu1_model:
          type: ["string", "null"]
        os:
          type: ["string", "null"]
        primary_ip:
          type: ["string", "null"]
        data:
          type: ["object", "null"]
          additionalProperties: true
        flags:
          type: array
          items:
            $ref: "#/components/schemas/MachineFlag"

    MachineReport:
      type: object
      required: [machine_id, hostname]
      properties:
        machine_id:
          type: string
        hostname:
          type: string
        site:
          type: ["string", "null"]
        manufacturer:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        cpu:
          type: ["string", "null"]
        total_ram_gb:
          type: ["integer", "null"]
        ram_type:
          type: ["string", "null"]
        gpu1_model:
          type: ["string", "null"]
        os:
          type: ["string", "null"]
        primary_ip:
          type: ["string", "null"]
        data:
          type: ["object", "null"]
          additionalProperties: true

    DashboardStats:
      type: object
      required: [total_machines, flagged_machines, site_count, danger_count, warn_count]
      properties:
        total_machines:
          type: integer
        flagged_machines:
          type: integer
        site_count:
          type: integer
        danger_count:
          type: integer
        warn_count:
          type: integer

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    UserSession:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string

    User:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string

    UserInput:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, viewer]

    PasswordReset:
      type: object
      required: [password]
      properties:
        password:
          type: string

    UpdateMachineSiteInput:
      type: object
      required: [site]
      properties:
        site:
          type: ["string", "null"]

    SubnetRule:
      type: object
      required: [id, cidr, site, created_at]
      properties:
        id:
          type: integer
        cidr:
          type: string
        site:
          type: string
        created_at:
          type: string
          format: date-time

    SubnetRuleInput:
      type: object
      required: [cidr, site]
      properties:
        cidr:
          type: string
          description: IPv4 CIDR, e.g. 10.1.0.0/16
        site:
          type: string