# lib/api-zod/src/generated/types/subnetRule.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fleet Inventory Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface SubnetRule {
  id: number;
  cidr: string;
  site: string;
  created_at: Date;
}