# lib/api-zod/src/generated/types/subnetRuleInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fleet Inventory Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface SubnetRuleInput {
  /** IPv4 CIDR, e.g. 10.1.0.0/16 */
  cidr: string;
  site: string;
}